requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
atlassian==0.0.0
atlassian-python-api==3.32.1
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
Deprecated==1.2.13
et-xmlfile==1.1.0
filelock==3.9.0
frozenlist==1.3.3
html2text==2020.1.16
huggingface-hub==0.12.0
idna==3.4
joblib==1.2.0
multidict==6.0.4
nltk==3.8
numpy==1.24.1
oauthlib==3.2.2
openai==0.25.0
openpyxl==3.0.10
packaging==23.0
pandas==1.5.2
pandas-stubs==1.5.2.230105
pdfminer.six==20221105
pycparser==2.21
pycryptodome==3.16.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
slackclient==2.9.4
soupsieve==2.3.2.post1
tokenizers==0.13.2
tqdm==4.64.1
transformers==4.25.1
types-pytz==2022.7.1.0
typing_extensions==4.4.0
urllib3==1.26.14
wrapt==1.14.1
yarl==1.8.2
